Move RequestParamNormalizer into controller
